app/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose
passlib[bcrypt]
python-dotenv
pydantic[email]
psycopg2  
python-multipart